move age histogram to its own file
